+++ lib/cisco_node_utils/platform.rb
@@ -188,7 +188,14 @@ module Cisco
     #                            'cpu'      => '0% system CPU' }},
     #      { ... }}
     def self.virtual_services
-      virts = config_get('virtual_service', 'services')
+      # If no virtual-services are installed, this will result
+      # in a RuntimeError.  We need to rescue this specific case.
+      begin
+        virts = config_get('virtual_service', 'services')
+      rescue RuntimeError => e
+        return {} if e.message[/No key \"TABLE_detail\"/]
+        raise
+      end
       return {} if virts.nil?
       # NXAPI returns hash instead of array if there's only 1
       virts = [virts] if virts.is_a? Hash